+++ gcc_kernel/__main__.py
@@ -1,0 +1,4 @@
+from ipykernel.kernelapp import IPKernelApp
+from . import GccKernel
+
+IPKernelApp.launch_instance(kernel_class=GccKernel)